Move `getBreadcrumbItems()` from `utils` to `Breadcrumb` component
